+++ CHANGELOG.md
@@ -656,3 +656,4 @@
 * Remove `Divariant` typeclass
 * Define `Turnover` wrapper for data structures that can that can be rotated indefinitely
 * Wrap left part of `Tape` into `Reverse`
+* Rename `-+-` operator to `.-+-` in Algebraic module